backend/foodgram_backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.4
defusedxml==0.8.0rc2
Django==3.2.3
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.6
mypy==1.8.0
mypy-extensions==1.0.0
oauthlib==3.2.2
pillow==10.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.2.1
